requirements.txt
flask
google-cloud-compute
gunicorn